+++ src/TSTransformer/nodes/binding/transformArrayBindingPattern.ts
@@ -1,11 +1,10 @@
 import luau from "@roblox-ts/luau-ast";
-import { errors } from "Shared/diagnostics";
 import { TransformState } from "TSTransformer";
-import { DiagnosticService } from "TSTransformer/classes/DiagnosticService";
 import { transformObjectBindingPattern } from "TSTransformer/nodes/binding/transformObjectBindingPattern";
 import { transformVariable } from "TSTransformer/nodes/statements/transformVariableStatement";
 import { transformInitializer } from "TSTransformer/nodes/transformInitializer";
 import { getAccessorForBindingType } from "TSTransformer/util/binding/getAccessorForBindingType";
+import { getDestructorForBindingType } from "TSTransformer/util/binding/getDestructorForBindingType";
 import { validateNotAnyType } from "TSTransformer/util/validateNotAny";
 import ts from "typescript";
 
@@ -18,17 +17,19 @@ export function transformArrayBindingPattern(
 
 	let index = 0;
 	const idStack = new Array<luau.AnyIdentifier>();
+
 	const accessor = getAccessorForBindingType(state, bindingPattern, state.getType(bindingPattern));
+	const destructor = getDestructorForBindingType(state, bindingPattern, state.getType(bindingPattern));
+
 	for (const element of bindingPattern.elements) {
 		if (ts.isOmittedExpression(element)) {
 			accessor(state, parentId, index, idStack, true);
 		} else {
-			if (element.dotDotDotToken) {
-				DiagnosticService.addDiagnostic(errors.noSpreadDestructuring(element));
-				return;
-			}
 			const name = element.name;
-			const value = accessor(state, parentId, index, idStack, false);
+			const value = element.dotDotDotToken
+				? destructor(state, parentId, index, idStack)
+				: accessor(state, parentId, index, idStack, false);
+
 			if (ts.isIdentifier(name)) {
 				const id = transformVariable(state, name, value);
 				if (element.initializer) {