+++ src/TSTransformer/nodes/binding/transformObjectBindingPattern.ts
@@ -1,11 +1,10 @@
 import luau from "@roblox-ts/luau-ast";
-import { errors } from "Shared/diagnostics";
 import { assert } from "Shared/util/assert";
 import { TransformState } from "TSTransformer";
-import { DiagnosticService } from "TSTransformer/classes/DiagnosticService";
 import { transformArrayBindingPattern } from "TSTransformer/nodes/binding/transformArrayBindingPattern";
 import { transformVariable } from "TSTransformer/nodes/statements/transformVariableStatement";
 import { transformInitializer } from "TSTransformer/nodes/transformInitializer";
+import { getDestructorForObjectType } from "TSTransformer/util/binding/getDestructorForObjectType";
 import { objectAccessor } from "TSTransformer/util/binding/objectAccessor";
 import { validateNotAnyType } from "TSTransformer/util/validateNotAny";
 import ts from "typescript";
@@ -16,15 +15,18 @@ export function transformObjectBindingPattern(
 	parentId: luau.AnyIdentifier,
 ) {
 	validateNotAnyType(state, bindingPattern);
+	const preSpreadNames = new Array<luau.Expression>();
 	for (const element of bindingPattern.elements) {
-		if (element.dotDotDotToken) {
-			DiagnosticService.addDiagnostic(errors.noSpreadDestructuring(element));
-			return;
-		}
 		const name = element.name;
 		const prop = element.propertyName;
+		const isSpread = element.dotDotDotToken !== undefined;
+
 		if (ts.isIdentifier(name)) {
-			const value = objectAccessor(state, parentId, state.getType(bindingPattern), prop ?? name);
+			const value = isSpread
+				? getDestructorForObjectType(state, parentId, preSpreadNames)
+				: objectAccessor(state, parentId, state.getType(bindingPattern), prop ?? name);
+			preSpreadNames.push(value);
+
 			const id = transformVariable(state, name, value);
 			if (element.initializer) {
 				state.prereq(transformInitializer(state, id, element.initializer));
@@ -33,7 +35,11 @@ export function transformObjectBindingPattern(
 			// if name is not identifier, it must be a binding pattern
 			// in that case, prop is guaranteed to exist
 			assert(prop);
-			const value = objectAccessor(state, parentId, state.getType(bindingPattern), prop);
+			const value = isSpread
+				? getDestructorForObjectType(state, parentId, preSpreadNames)
+				: objectAccessor(state, parentId, state.getType(bindingPattern), prop);
+			preSpreadNames.push(value);
+
 			const id = state.pushToVar(value, "binding");
 			if (element.initializer) {
 				state.prereq(transformInitializer(state, id, element.initializer));