+++ src/TSTransformer/nodes/transformParameters.ts
@@ -6,6 +6,7 @@ import { transformArrayBindingPattern } from "TSTransformer/nodes/binding/transf
 import { transformObjectBindingPattern } from "TSTransformer/nodes/binding/transformObjectBindingPattern";
 import { transformIdentifierDefined } from "TSTransformer/nodes/expressions/transformIdentifier";
 import { transformInitializer } from "TSTransformer/nodes/transformInitializer";
+import { arrayLikeExpressionContainsSpread } from "TSTransformer/util/arrayBindingPatternContainsSpread";
 import { isMethod } from "TSTransformer/util/isMethod";
 import { validateIdentifier } from "TSTransformer/util/validateIdentifier";
 import ts from "typescript";
@@ -22,10 +23,10 @@ function optimizeArraySpreadParameter(
 		if (ts.isOmittedExpression(element)) {
 			luau.list.push(parameters, luau.tempId());
 		} else {
-			if (element.dotDotDotToken) {
-				DiagnosticService.addDiagnostic(errors.noSpreadDestructuring(element));
-				return;
-			}
+			// if (element.dotDotDotToken) {
+			// 	DiagnosticService.addDiagnostic(errors.noSpreadDestructuring(element));
+			// 	return;
+			// }
 			const name = element.name;
 			if (ts.isIdentifier(name)) {
 				const paramId = transformIdentifierDefined(state, name);
@@ -64,7 +65,11 @@ export function transformParameters(state: TransformState, node: ts.SignatureDec
 			continue;
 		}
 
-		if (parameter.dotDotDotToken && ts.isArrayBindingPattern(parameter.name)) {
+		if (
+			parameter.dotDotDotToken &&
+			ts.isArrayBindingPattern(parameter.name) &&
+			!arrayLikeExpressionContainsSpread(parameter.name)
+		) {
 			const prereqs = state.capturePrereqs(() =>
 				optimizeArraySpreadParameter(state, parameters, parameter.name as ts.ArrayBindingPattern),
 			);